compression_aware_forensics/requirements.txt
Flask==2.3.3
torch==2.9.0
torchvision==0.24.0
Pillow==10.0.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.0.76
scipy==1.11.3
